refactor: allow to use BillingService model outside of autorenew module
