+++ client/app/billing/autoRenew/actions/update/update.controller.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import BillingService from '../../BillingService.class';
+import BillingService from '../../../../models/BillingService.class';
 
 export default class {
   /* @ngInject */
